backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
openpyxl
gunicorn
resend
python-dotenv
